+++ src/Products/PlonePAS/tests/test_memberdatatool.py
@@ -1,9 +1,13 @@
 # -*- coding: utf-8 -*-
 from DateTime import DateTime
 from OFS.Image import Image
+from Products.CMFCore.interfaces import IMemberData
 from Products.PlonePAS.tests import base
 from Products.PlonePAS.tests import dummy
+from Products.PluggableAuthService.interfaces.events import \
+        IPropertiesUpdatedEvent
 from plone.app.testing import TEST_USER_ID as default_user
+import zope.component
 
 
 class TestMemberDataTool(base.TestCase):
@@ -68,3 +72,17 @@ class TestMemberDataTool(base.TestCase):
         self.assertEqual(len(search('bambam.net')), 1)
         self.assertEqual(len(search('bedrock.com')), 2)
         self.assertEqual(len(search('brubble')), 1)
+
+    def testPropertiesUpdatedEvent(self):
+
+        def event_handler(context, event):
+            self._properties_updated_handler_called = True
+
+        gsm = zope.component.getGlobalSiteManager()
+        gsm.registerHandler(event_handler,
+                            (IMemberData, IPropertiesUpdatedEvent))
+
+        self._properties_updated_handler_called = False
+        self.addMember('ez', 'Ez Zy', 'ez@ezmail.net',
+                       ['Member'], '2018-01-01')
+        self.assertTrue(self._properties_updated_handler_called)